refactor: CoordInput を1つに統一
